Move the MoviesDetails component inot the container directory
