+++ src/actions/constants.js
@@ -1,0 +1,5 @@
+export const SEARCH_MOVIE = 'SEARCH_MOVIE';
+export const FETCH_MOVIES = 'FETCH_MOVIES';
+export const FETCH_MOVIE_DETAILS = 'FETCH_MOVIE_DETAILS';
+export const LOADING = 'LOADING';
+export const FILTER_MOVIE = 'FILTER_MOVIE';
